requirements.txt
colorama==0.4.6
dotenv==0.9.9
numpy==2.3.3
pandas==2.3.2
psycopg==3.2.10
psycopg-binary==3.2.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
six==1.17.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
